+++ public_html/webmaster/settings/aboutpage.php
@@ -1,0 +1,5 @@
+<?php
+$title = "About Page";
+$slug = "about";
+include ".page.php";
+?>