+++ public_html/webmaster/settings/legacy.php
@@ -1,5 +1,5 @@
 <?php
-require_once "../../ArtArchive.php";
+require_once "../../../ArtArchive.php";
 ArtArchive::RequireWebmaster();
 
 $bdd = &ArtArchive::$database;