+++ changelog.md
@@ -56,3 +56,4 @@
 ## 0.9-indev
 - Added hotlist.php
 - Expanded the Error Document a little
+- Reorganized setting page into multiple pages